diff_two_coordinates and is_nearby_danger
I noticed that I would be using diff_two_coordinates in two different
spots of the player.rs file, so I made a private function for the Player
Struct.

Also, I was able to calculate nearby danger to warn the player.  The
Player still doesn't lose a life though if he hits a danger spot!